requirements.txt
Django==2.2.1
django-constance==2.4.0
django-phone-field==1.7.1
django-picklefield==2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
Pillow==6.0.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
